Transformer refactor, Logic from Endpoints to service refactor, postman collections added

+++ maps/backend/src/main/java/com/globant/training/google/maps/antenna/service/AntennaServiceImpl.java
@@ -1,9 +1,13 @@
 package com.globant.training.google.maps.antenna.service;
 
 import com.google.inject.Inject;
+
 import com.globant.training.google.maps.antenna.dao.AntennaDao;
 import com.globant.training.google.maps.antenna.entity.Antenna;
 
+import org.apache.commons.lang3.Validate;
+
+import java.util.Date;
 import java.util.List;
 
 /**
@@ -20,44 +24,58 @@ public class AntennaServiceImpl implements AntennaService {
     super();
     this.antennaDao = antennaDao;
   }
-
-  /*
-   * (non-Javadoc)
-   * 
-   * @see com.globant.training.google.maps.services.AntennaService#getAll()
-   */
+  
   @Override
   public List<Antenna> getAll() {
     return antennaDao.getAll();
   }
 
-  /*
-   * (non-Javadoc)
-   * 
-   * @see
-   * com.globant.training.google.maps.services.AntennaService#save(com.globant.training.google.maps.
-   * entities.Antenna)
-   */
   @Override
-  public Antenna save(Antenna antenna) {
+  public Antenna create(Antenna antenna) {
+    Validate.notNull(antenna, "Antenna cannot be null");
+
+    Date now = new Date();
+    antenna.setCreated(now);
+    antenna.setLastUpdated(now);
+
     return antennaDao.put(antenna);
   }
 
-  /*
-   * (non-Javadoc)
-   * 
-   * @see com.globant.training.google.maps.services.AntennaService#findById(java.lang.Long)
-   */
+  @Override
+  public Antenna update(Long id, Antenna antenna) {
+    Validate.notNull(antenna, "id cannot be null");
+    Validate.notNull(antenna, "Antenna cannot be null");
+
+    Antenna existingAntenna = findById(id);
+
+    if (existingAntenna == null) {
+      throw new RuntimeException("Antenna Not Found");
+    }
+
+    antenna.setId(existingAntenna.getId());
+    antenna.setCreated(existingAntenna.getCreated());
+    antenna.setLastUpdated(new Date());
+
+    return antennaDao.put(antenna);
+  }
+
   @Override
   public Antenna findById(Long id) {
+    Validate.notNull(id, "id cannot be null");
+
     return antennaDao.get(id);
   }
 
-  /* (non-Javadoc)
-   * @see com.globant.training.google.maps.services.AntennaService#deleteById(java.lang.Long)
-   */
   @Override
   public void deleteById(Long id) {
+    Validate.notNull(id, "id cannot be null");
+
+    Antenna antenna = findById(id);
+
+    if (antenna == null) {
+      throw new RuntimeException("Antenna Not Found");
+    }
+
     antennaDao.delete(id);
   }
